web/requirements.txt
streamlit>=1.28.0
gspread>=5.11.0
google-auth>=2.23.0
pandas>=2.0.0
streamlit-autorefresh>=0.0.1
python-dateutil>=2.8.2
pytz>=2023.3